requirements.txt
langchain==0.0.320
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.27.2
huggingface-hub==0.18.0
InstructorEmbedding==1.0.1
faiss-cpu==1.7.4
tiktoken==0.5.1
sentence-transformer==2.2.2
transformers==4.34.1